Fix terrain data management
